AI_Employee_Vault/silver/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-employee-silver"
version = "1.0.0"
description = "Silver Tier - Functional AI Assistant with multi-channel monitoring and HITL approval"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "AI Employee Vault", email = "noreply@example.com"}
]
keywords = ["ai", "assistant", "automation", "obsidian", "gmail", "whatsapp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Communications :: Email",
]

dependencies = [
    # Core dependencies (from Bronze tier)
    "watchdog>=4.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",

    # Gmail API
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.115.0",

    # WhatsApp Web automation
    "playwright>=1.40.0",

    # Desktop notifications
    "plyer>=2.1.0",

    # Scheduling
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-employee-vault"
Documentation = "https://github.com/yourusername/ai-employee-vault/tree/main/silver"
Repository = "https://github.com/yourusername/ai-employee-vault"
Issues = "https://github.com/yourusername/ai-employee-vault/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py313"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
